refactor: Service 분리
[- Command, Query로 분리]
