+++ src/main/java/com/watermelon/server/Scheduler.java
@@ -1,11 +1,10 @@
 package com.watermelon.server;
 
-import com.watermelon.server.fifoevent.service.OrderEventService;
+import com.watermelon.server.orderevent.service.OrderEventCommandService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 
 
 @Slf4j
@@ -13,11 +12,11 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class Scheduler {
 
-    private final OrderEventService orderEventService;
+    private final OrderEventCommandService orderEventCommandService;
 
     @Scheduled(fixedRate = 1000)
     public void checkEventStart(){
         log.info("Checking events by scheduled");
-        orderEventService.changeOrderStatusByTime();
+        orderEventCommandService.changeOrderStatusByTime();
     }
 }
