--- src/main/java/com/watermelon/server/fifoevent/domain/OrderEventStatus.java
@@ -1,6 +1,0 @@
-package com.watermelon.server.fifoevent.domain;
-
-public enum OrderEventStatus {
-
-    UPCOMING,OPEN,CLOSED,END;
-}
