--- src/main/java/com/watermelon/server/fifoevent/domain/OrderEventWinner.java
@@ -1,17 +1,0 @@
-package com.watermelon.server.fifoevent.domain;
-
-
-import jakarta.persistence.*;
-
-@Entity
-public class OrderEventWinner {
-    @Id @GeneratedValue
-    private Long id;
-
-    private String phoneNumber;
-
-    @ManyToOne
-    @JoinColumn
-    private OrderEvent orderEvent;
-
-}
