--- src/main/java/com/watermelon/server/fifoevent/dto/request/RequestAnswerDto.java
@@ -1,16 +1,0 @@
-package com.watermelon.server.fifoevent.dto.request;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class RequestAnswerDto {
-    private String answer;
-    private Long fifoEventId;
-}
