--- src/main/java/com/watermelon/server/fifoevent/dto/request/RequestQuizDto.java
@@ -1,16 +1,0 @@
-package com.watermelon.server.fifoevent.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class RequestQuizDto {
-
-    private String description;
-    private String answer;
-    private String imgSrc;
-    private String title;
-}
