--- src/main/java/com/watermelon/server/fifoevent/repository/QuizRepository.java
@@ -1,9 +1,0 @@
-package com.watermelon.server.fifoevent.repository;
-
-import com.watermelon.server.fifoevent.domain.Quiz;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface QuizRepository extends JpaRepository<Quiz,Long> {
-}
