--- src/main/java/com/watermelon/server/fifoevent/service/OrderEventService.java
@@ -1,81 +1,0 @@
-package com.watermelon.server.fifoevent.service;
-
-
-import com.watermelon.server.fifoevent.domain.OrderEvent;
-import com.watermelon.server.fifoevent.domain.Quiz;
-import com.watermelon.server.fifoevent.dto.request.RequestAnswerDto;
-import com.watermelon.server.fifoevent.dto.request.RequestOrderEventDto;
-import com.watermelon.server.fifoevent.dto.response.ResponseOrderEventDto;
-import com.watermelon.server.fifoevent.dto.response.ResponseQuizResultDto;
-import com.watermelon.server.fifoevent.repository.OrderEventRepository;
-import com.watermelon.server.fifoevent.repository.QuizRepository;
-import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-@Service
-@RequiredArgsConstructor
-public class OrderEventService {
-    private static final Logger log = LoggerFactory.getLogger(OrderEventService.class);
-    private final OrderEventRepository orderEventRepository;
-    private final QuizRepository quizRepository;
-
-
-    public List<ResponseOrderEventDto> getOrderEvents(){
-        List<OrderEvent> orderEvents = orderEventRepository.findAll();
-        List<ResponseOrderEventDto> responseOrderEventDtos = new ArrayList<>();
-        orderEvents.forEach(orderEvent -> responseOrderEventDtos.add(ResponseOrderEventDto.from(orderEvent)));
-        return responseOrderEventDtos;
-    }
-
-    @Transactional
-    public void changeOrderStatusByTime(){
-        List<OrderEvent> orderEvents = orderEventRepository.findAll();
-        orderEvents.forEach(orderEvent -> {orderEvent.changeOrderEventStatusByTime(LocalDateTime.now());});
-    }
-
-    @Transactional
-    public OrderEvent makeEvent(RequestOrderEventDto requestOrderEventDto){
-        OrderEvent newOrderEvent = OrderEvent.makeOrderEvent(requestOrderEventDto);
-        return orderEventRepository.save(newOrderEvent);
-    }
-
-
-
-    public ResponseQuizResultDto applyFifoEvent(RequestAnswerDto requestAnswerDto){
-        Optional<OrderEvent> fifoEventOptional = orderEventRepository.findById(requestAnswerDto.getFifoEventId());
-        OrderEvent orderEvent = fifoEventOptional.get();
-
-        // 시간 틀림
-        if(!orderEvent.isTimeInEventTime(LocalDateTime.now())){
-            return ResponseQuizResultDto.eventTimeOutResult();
-        }
-        Quiz quiz = orderEvent.getQuiz();
-
-
-        //정답 틀림
-        if(!quiz.isCorrect(requestAnswerDto.getAnswer())){
-            return ResponseQuizResultDto.wrongAnswerResult();
-        }
-
-        //선착순 틀림
-        if(!orderEvent.isWinnerAddable()){
-            return ResponseQuizResultDto.noMoreWinnerResult();
-        }
-
-        //선착순 토큰 발급
-        orderEvent.addWinner();
-        return ResponseQuizResultDto.answerRightResult(UUID.randomUUID().toString());
-
-
-    }
-
-}
