+++ src/main/java/com/watermelon/server/orderevent/domain/OrderEventStatus.java
@@ -1,0 +1,6 @@
+package com.watermelon.server.orderevent.domain;
+
+public enum OrderEventStatus {
+
+    UPCOMING,OPEN,CLOSED,END;
+}
