+++ src/main/java/com/watermelon/server/orderevent/domain/OrderEventWinner.java
@@ -1,0 +1,17 @@
+package com.watermelon.server.orderevent.domain;
+
+
+import jakarta.persistence.*;
+
+@Entity
+public class OrderEventWinner {
+    @Id @GeneratedValue
+    private Long id;
+
+    private String phoneNumber;
+
+    @ManyToOne
+    @JoinColumn
+    private OrderEvent orderEvent;
+
+}
