+++ src/main/java/com/watermelon/server/orderevent/dto/request/RequestAnswerDto.java
@@ -1,0 +1,16 @@
+package com.watermelon.server.orderevent.dto.request;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class RequestAnswerDto {
+    private String answer;
+    private Long fifoEventId;
+}
