+++ src/main/java/com/watermelon/server/orderevent/dto/request/RequestOrderEventDto.java
@@ -1,0 +1,22 @@
+package com.watermelon.server.orderevent.dto.request;
+
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class RequestOrderEventDto {
+    private LocalDateTime startTime;
+    private LocalDateTime endTime;
+    private int maxWinnerCount;
+
+    private RequestQuizDto requestQuizDto;
+    private RequestOrderRewardDto requestOrderRewardDto;
+
+
+
+
+}
