+++ src/main/java/com/watermelon/server/orderevent/dto/request/RequestQuizDto.java
@@ -1,0 +1,16 @@
+package com.watermelon.server.orderevent.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class RequestQuizDto {
+
+    private String description;
+    private String answer;
+    private String imgSrc;
+    private String title;
+}
