+++ src/main/java/com/watermelon/server/orderevent/dto/response/ResponseQuizResultSuccessDto.java
@@ -1,0 +1,14 @@
+package com.watermelon.server.orderevent.dto.response;
+
+
+import lombok.Getter;
+
+@Getter
+public class ResponseQuizResultSuccessDto extends  ResponseQuizResultDto {
+    private String applyToken;
+
+    public ResponseQuizResultSuccessDto(boolean isSuccess, String applyToken) {
+        super(isSuccess);
+        this.applyToken = applyToken;
+    }
+}
