+++ src/main/java/com/watermelon/server/orderevent/repository/QuizRepository.java
@@ -1,0 +1,9 @@
+package com.watermelon.server.orderevent.repository;
+
+import com.watermelon.server.orderevent.domain.Quiz;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface QuizRepository extends JpaRepository<Quiz,Long> {
+}
